cleaned up cli for better readability
